% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/location.R
\name{orderly_location_azure}
\alias{orderly_location_azure}
\title{Orderly Location for Azure file storage}
\description{
A driver to put orderly2 archives on
Azure. Individual methods are not documented here as this
just needs to satisfy the interface required by orderly2
and is not for direct user use. See
\link[orderly2:orderly_location_add]{orderly2::orderly_location_add} for details.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-orderly_location_azure-new}{\code{orderly_location_azure$new()}}
\item \href{#method-orderly_location_azure-verify}{\code{orderly_location_azure$verify()}}
\item \href{#method-orderly_location_azure-list}{\code{orderly_location_azure$list()}}
\item \href{#method-orderly_location_azure-metadata}{\code{orderly_location_azure$metadata()}}
\item \href{#method-orderly_location_azure-fetch_file}{\code{orderly_location_azure$fetch_file()}}
\item \href{#method-orderly_location_azure-list_unknown_packets}{\code{orderly_location_azure$list_unknown_packets()}}
\item \href{#method-orderly_location_azure-list_unknown_files}{\code{orderly_location_azure$list_unknown_files()}}
\item \href{#method-orderly_location_azure-push_file}{\code{orderly_location_azure$push_file()}}
\item \href{#method-orderly_location_azure-push_metadata}{\code{orderly_location_azure$push_metadata()}}
\item \href{#method-orderly_location_azure-clone}{\code{orderly_location_azure$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-orderly_location_azure-new"></a>}}
\if{latex}{\out{\hypertarget{method-orderly_location_azure-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{orderly_location_azure$new(url, container, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{url}}{The base url of your Azure file share, such
as \verb{https://storename.file.core.windows.net/}
(the \verb{https://} prefix is required)}

\item{\code{container}}{Name of the storage container on Azure i.e. name of
the file share}

\item{\code{...}}{Additional args for authenticating with Azure, passed
straight through to \link[AzureStor:storage_endpoint]{AzureStor::storage_endpoint}}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-orderly_location_azure-verify"></a>}}
\if{latex}{\out{\hypertarget{method-orderly_location_azure-verify}{}}}
\subsection{Method \code{verify()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{orderly_location_azure$verify()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-orderly_location_azure-list"></a>}}
\if{latex}{\out{\hypertarget{method-orderly_location_azure-list}{}}}
\subsection{Method \code{list()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{orderly_location_azure$list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-orderly_location_azure-metadata"></a>}}
\if{latex}{\out{\hypertarget{method-orderly_location_azure-metadata}{}}}
\subsection{Method \code{metadata()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{orderly_location_azure$metadata(packet_ids)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-orderly_location_azure-fetch_file"></a>}}
\if{latex}{\out{\hypertarget{method-orderly_location_azure-fetch_file}{}}}
\subsection{Method \code{fetch_file()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{orderly_location_azure$fetch_file(hash, dest)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-orderly_location_azure-list_unknown_packets"></a>}}
\if{latex}{\out{\hypertarget{method-orderly_location_azure-list_unknown_packets}{}}}
\subsection{Method \code{list_unknown_packets()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{orderly_location_azure$list_unknown_packets(ids)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-orderly_location_azure-list_unknown_files"></a>}}
\if{latex}{\out{\hypertarget{method-orderly_location_azure-list_unknown_files}{}}}
\subsection{Method \code{list_unknown_files()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{orderly_location_azure$list_unknown_files(hashes)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-orderly_location_azure-push_file"></a>}}
\if{latex}{\out{\hypertarget{method-orderly_location_azure-push_file}{}}}
\subsection{Method \code{push_file()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{orderly_location_azure$push_file(src, hash)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-orderly_location_azure-push_metadata"></a>}}
\if{latex}{\out{\hypertarget{method-orderly_location_azure-push_metadata}{}}}
\subsection{Method \code{push_metadata()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{orderly_location_azure$push_metadata(packet_id, hash, path)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-orderly_location_azure-clone"></a>}}
\if{latex}{\out{\hypertarget{method-orderly_location_azure-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{orderly_location_azure$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
